Simplify predicate evaluator module hierarchy
